chore(travis): split tests into unit and integration tests
Coverage should be calculated from only unit tests, not integration tests.
