+++ .travis.yml
@@ -2,14 +2,21 @@ language: node_js
 node_js:
 - '0.12'
 sudo: false
+env:
+- TEST_SUITE=units
+- TEST_SUITE=integration
+before_install:
+- npm run install-bats
+before_script:
+- npm run compile
+script:
+- npm run test-$TEST_SUITE
+after_success:
+- npm run codecov
 cache:
   directories:
   - node_modules
   - test/modules
-before_install:
-- npm run install-bats
-after_success:
-- npm run codecov
 deploy:
   provider: npm
   email: git@jwr.vc
