+++ package.json
@@ -4,11 +4,12 @@
   "description": "Markdown transcluder",
   "main": "./lib/hercule",
   "scripts": {
-    "test": "npm run compile && npm run unit-test && npm run integration-test",
+    "test": "npm run compile && npm run test-units && npm run test-integration && npm run test-cli",
+    "test-units": "./node_modules/.bin/nyc --reporter=json ./node_modules/.bin/ava",
+    "test-integration": "./node_modules/.bin/ava --serial test/integration",
+    "test-cli": "./test/modules/bin/bats test/bats",
     "posttest": "npm run lint",
-    "unit-test": "./node_modules/.bin/nyc --reporter=json ./node_modules/.bin/ava",
-    "integration-test": "./test/modules/bin/bats test/bats",
-    "coverage": "./scripts/coverage && ./node_modules/.bin/istanbul report --reporter=html && open coverage/index.html",
+    "coverage": "./scripts/coverage && ./node_modules/.bin/istanbul report --reporter=html && open coverage/lcov-report/index.html",
     "codecov": "./scripts/coverage && cat coverage/coverage.json | ./node_modules/.bin/codecov",
     "compile": "BABEL_ENV=compile ./node_modules/.bin/babel src --out-dir lib --source-maps && ./node_modules/.bin/pegjs src/transclude.pegjs lib/transclude-parser.js",
     "install-bats": "./scripts/install-bats",
