+++ test/integration/transcludeStringSync.js
@@ -1,15 +1,19 @@
 import test from 'ava';
+import path from 'path';
 import _ from 'lodash';
 
-import {transcludeFileSync} from '../lib/hercule';
-import fixtures from './fixtures';
+import {transcludeStringSync} from '../../lib/hercule';
+import fixtures from '../fixtures';
 
 
 _.forEach((fixtures.fixtures), (fixture) => {
   // Exclude http test because mocking won't cover sync sub-process
   if (fixture.name !== 'http-link') {
     test(`should transclude ${fixture.name}`, (t) => {
-      const output = transcludeFileSync(fixture.inputFile);
+      const options = {
+        relativePath: path.resolve(__dirname, '../fixtures', fixture.name),
+      };
+      const output = transcludeStringSync(fixture.input, options);
       t.same(output, fixture.expectedOutput);
     });
   }
